+++ modules/apps/static/portal-lpkg-deployer/portal-lpkg-deployer-impl/src/main/java/com/liferay/portal/lpkg/deployer/internal/osgi/util/tracker/URLStreamHandlerServiceServiceTrackerCustomizer.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.lpkg.deployer.internal.wrapper.bundle;
+package com.liferay.portal.lpkg.deployer.internal.osgi.util.tracker;
 
 import com.liferay.osgi.util.bundle.BundleStartLevelUtil;
 import com.liferay.portal.lpkg.deployer.internal.WABWrapperUtil;
